Refactor components out of app.jsx
The remaining components which still live in app.jsx are refactored out into the components/ dir. Also, fix app.jsx indents to match other files (2 spaces per indent).
